Improved data types
Gotta save those extra bytes.

+++ Persistence/database/accounts/structure.sql
@@ -6,18 +6,18 @@ CREATE TABLE `accounts` (
     `id`                   int          NOT NULL AUTO_INCREMENT             COMMENT 'Primary Key.',
     `users_id`             int          NOT NULL                            COMMENT 'Account''s owner.',
     `session_id`           varchar(32)  NOT NULL                            COMMENT 'Session ID.',
-    `levels_id`            int          NOT NULL DEFAULT 1                  COMMENT 'Current level.',
-    `factions_id`          int          NULL     DEFAULT NULL               COMMENT 'Faction that the account belongs to.',
+    `levels_id`            tinyint      NOT NULL DEFAULT 1                  COMMENT 'Current level.',
+    `factions_id`          tinyint      NULL     DEFAULT NULL               COMMENT 'Faction that the account belongs to.',
     `accounts_hangars_id`  int          NULL     DEFAULT NULL               COMMENT 'Active hangar.',
     `clans_id`             int          NULL     DEFAULT NULL,
-    `ranks_id`             int          NOT NULL DEFAULT 1,
+    `ranks_id`             tinyint      NOT NULL DEFAULT 1,
     `name`                 varchar(255) NOT NULL DEFAULT ''                 COMMENT 'In game name.',
     `ban_date`             timestamp    NULL     DEFAULT NULL               COMMENT 'Ban expiration date.',
     `premium_date`         timestamp    NULL     DEFAULT NULL               COMMENT 'Premium expiration date.',
     `date`                 timestamp    NOT NULL DEFAULT CURRENT_TIMESTAMP,
     `last_login`           timestamp    NULL     DEFAULT NULL               COMMENT 'Last login date.',
-    `skill_points_total`   int          NOT NULL DEFAULT 0                  COMMENT 'Total skill points available.',
-    `skill_points_free`    int          NOT NULL DEFAULT 0                  COMMENT 'Free skill points available.',
+    `skill_points_total`   smallint     NOT NULL DEFAULT 0                  COMMENT 'Total skill points available.',
+    `skill_points_free`    smallint     NOT NULL DEFAULT 0                  COMMENT 'Free skill points available.',
 
     CONSTRAINT `accounts_pk` PRIMARY KEY (`id`)
 ) ENGINE InnoDB CHARACTER SET utf8 COMMENT 'In game accounts.';
