+++ Persistence/database/accounts_drones/structure.sql
@@ -5,8 +5,8 @@
 CREATE TABLE `accounts_drones` (
     `id`          int       NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
     `accounts_id` int       NOT NULL,
-    `levels_id`   int       NOT NULL DEFAULT 1,
-    `experience`  int       NOT NULL DEFAULT 0,
+    `levels_id`   tinyint   NOT NULL DEFAULT 1,
+    `experience`  smallint  NOT NULL DEFAULT 0,
     `date`        timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
 
     CONSTRAINT `accounts_drones_pk` PRIMARY KEY (`id`)
