+++ Persistence/database/accounts_techfactory_items/structure.sql
@@ -5,8 +5,8 @@
 CREATE TABLE `accounts_techfactory_items` (
     `id`                   int       NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
     `accounts_id`          int       NOT NULL                COMMENT 'Account ID.',
-    `techfactory_items_id` int       NOT NULL                COMMENT 'Item ID.',
-    `amount`               int       NOT NULL DEFAULT 1      COMMENT 'Amount of build items.',
+    `techfactory_items_id` tinyint   NOT NULL                COMMENT 'Item ID.',
+    `amount`               smallint  NOT NULL DEFAULT 1      COMMENT 'Amount of build items.',
     `date`                 timestamp NULL     DEFAULT NULL   COMMENT 'Date when the item started building.',
 
     CONSTRAINT `accounts_techfactory_items_pk` PRIMARY KEY (`id`)
