+++ Persistence/database/clans_banks/structure.sql
@@ -3,12 +3,12 @@
 -- Clan's internal bank.
 --
 CREATE TABLE `clans_banks` (
-    `id`          int          NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
-    `clans_id`    int          NOT NULL,
-    `credits`     bigint       NOT NULL DEFAULT 0      COMMENT 'Credits available in the bank.',
-    `uridium`     bigint       NOT NULL DEFAULT 0      COMMENT 'Uridium available in the bank.',
-    `tax_credits` decimal(5,2) NOT NULL DEFAULT 5.0    COMMENT 'Tax rate for credits.',
-    `tax_uridium` decimal(5,2) NOT NULL DEFAULT 0.0    COMMENT 'Tax rate for uridium.',
+    `id`          int     NOT NULL AUTO_INCREMENT COMMENT 'Primary Key.',
+    `clans_id`    int     NOT NULL,
+    `credits`     bigint  NOT NULL DEFAULT 0      COMMENT 'Credits available in the bank.',
+    `uridium`     bigint  NOT NULL DEFAULT 0      COMMENT 'Uridium available in the bank.',
+    `tax_credits` tinyint NOT NULL DEFAULT 5.0    COMMENT 'Tax rate for credits.',
+    `tax_uridium` tinyint NOT NULL DEFAULT 0.0    COMMENT 'Tax rate for uridium.',
 
     CONSTRAINT `clans_banks_pk` PRIMARY KEY (`id`)
 ) ENGINE InnoDB CHARACTER SET utf8 COMMENT 'Clan''s internal bank.';
