+++ Persistence/database/clans_battlestations_items/structure.sql
@@ -3,11 +3,11 @@
 -- Items equipped in the CBS.
 --
 CREATE TABLE `clans_battlestations_items` (
-    `id`                      int        NOT NULL AUTO_INCREMENT            COMMENT 'Primary Key.',
-    `clans_battlestations_id` int        NOT NULL                           COMMENT 'CBS where item is equipped.',
-    `accounts_items_id`       int        NOT NULL                           COMMENT 'Equipped item.',
-    `slot`                    tinyint(1) NOT NULL DEFAULT 1                 COMMENT 'Position where the item is equipped (A = 9, B = 10).',
-    `date`                    timestamp  NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Date when the item was equipped.',
+    `id`                      int       NOT NULL AUTO_INCREMENT            COMMENT 'Primary Key.',
+    `clans_battlestations_id` tinyint   NOT NULL                           COMMENT 'CBS where item is equipped.',
+    `accounts_items_id`       int       NOT NULL                           COMMENT 'Equipped item.',
+    `slot`                    tinyint   NOT NULL DEFAULT 1                 COMMENT 'Position where the item is equipped (A = 9, B = 10).',
+    `date`                    timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Date when the item was equipped.',
 
     CONSTRAINT `clans_battlestations_items_pk` PRIMARY KEY (`id`)
 ) ENGINE InnoDB CHARACTER SET utf8 COMMENT 'Items equipped in the CBS.';
