+++ Persistence/database/clans_battlestations_logs/structure.sql
@@ -5,7 +5,7 @@
 CREATE TABLE `clans_battlestations_logs` (
     `id`                      int       NOT NULL AUTO_INCREMENT             COMMENT 'Primary Key.',
     `clans_id`                int       NOT NULL,
-    `clans_battlestations_id` int       NOT NULL,
+    `clans_battlestations_id` tinyint   NOT NULL,
     `message`                 text      NOT NULL,
     `date`                    timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
